storage/requirements.txt
boto3==1.18.21
confluent-kafka==1.7.0
pyarrow==5.0.0
PyYAML==5.4.1
